requirements.txt
dreamtools-dreamgeeker==1.3.5
Eel==0.16.0
libsass==0.22.0
Pillow==9.5.0
pyinstaller==5.10.1
qrcode==7.4.2
rcssmin==1.1.1
rjsmin==1.2.1
segno==1.5.2
six==1.16.0
soupsieve==2.4.1
typing_extensions==4.5.0
urllib3==1.26.15
validator==0.7.1
verify-email==2.4.3
whichcraft==0.6.1
zope.event==4.6
zope.interface==6.0
